fix uint80s

+++ contracts/GlobalsAndUtility.sol
@@ -172,8 +172,8 @@ contract GlobalsAndUtility is ERC20 {
     struct GlobalsStore {
         // 1
         uint16 daysStored;
-        uint80 stakeSharesTotal;
-        uint80 nextStakeSharesTotal;
+        uint256 stakeSharesTotal;
+        uint256 nextStakeSharesTotal;
         uint256 sharesPerHeart;
         uint48 latestStakeId;
         // 2
@@ -191,7 +191,7 @@ contract GlobalsAndUtility is ERC20 {
     /* Period data */
     struct DailyDataStore {
         uint80 dayPayoutTotal;
-        uint80 dayStakeSharesTotal;
+        uint256 dayStakeSharesTotal;
     }
 
     mapping(uint256 => DailyDataStore) public dailyData;
@@ -209,7 +209,7 @@ contract GlobalsAndUtility is ERC20 {
     struct StakeStore {
         uint48 stakeId;
         uint80 stakedHearts;
-        uint80 stakeShares;
+        uint256 stakeShares;
         uint16 pooledDay;
         uint16 stakedDays;
         uint16 unpooledDay;
@@ -382,8 +382,8 @@ contract GlobalsAndUtility is ERC20 {
         internal
     {
         globals.daysStored = uint16(g._daysStored);
-        globals.stakeSharesTotal = uint80(g._stakeSharesTotal);
-        globals.nextStakeSharesTotal = uint80(g._nextStakeSharesTotal);
+        globals.stakeSharesTotal = g._stakeSharesTotal;
+        globals.nextStakeSharesTotal = g._nextStakeSharesTotal;
         globals.sharesPerHeart = g._sharesPerHeart;
         globals.latestStakeId = g._latestStakeId;
     }
@@ -442,7 +442,7 @@ contract GlobalsAndUtility is ERC20 {
     {
         stRef.stakeId = st._stakeId;
         stRef.stakedHearts = uint80(st._stakedHearts);
-        stRef.stakeShares = uint80(st._stakeShares);
+        stRef.stakeShares = st._stakeShares;
         stRef.pooledDay = uint16(st._pooledDay);
         stRef.stakedDays = uint16(st._stakedDays);
         stRef.unpooledDay = uint16(st._unpooledDay);
@@ -462,7 +462,7 @@ contract GlobalsAndUtility is ERC20 {
             StakeStore(
                 newStakeId,
                 uint80(newStakedHearts),
-                uint80(newStakeShares),
+                newStakeShares,
                 uint16(newPooledDay),
                 uint16(newStakedDays),
                 uint16(0) // unpooledDay
@@ -529,7 +529,7 @@ contract GlobalsAndUtility is ERC20 {
             if (g._stakeSharesTotal != 0) {
                 _calcDailyRound(g, rs, day);
                 dailyData[day].dayPayoutTotal = uint80(rs._payoutTotal);
-                dailyData[day].dayStakeSharesTotal = uint80(g._stakeSharesTotal);
+                dailyData[day].dayStakeSharesTotal = g._stakeSharesTotal;
             } else {
                 if (day == CLAIM_REWARD_DAYS && g._unclaimedSatoshisTotal != 0) {
                     /*
